Skip Intercepted Header test for go <1.6
